+++ system/src/handlebars.mjs
@@ -230,4 +230,11 @@ export default function registerHandlebarsHelpers() {
 		return str;
 	});
 
+	Handlebars.registerHelper("select", function(selected, options) {
+		const escapedValue = RegExp.escape(Handlebars.escapeExpression(selected));
+		const rgx = new RegExp(` value=["']${escapedValue}["']`);
+		const html = options.fn(this);
+		return html.replace(rgx, "$& selected");
+	});
+
 }